Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/rendering/dark-color-scheme/color-scheme-color-property-expected.txt


FAIL Check that color:initial resolves to the intial color for the dark color-scheme. assert_equals: expected "rgb(255, 255, 255)" but got "rgb(0, 0, 0)"
FAIL Check that color:initial resolves to the intial color for the light color-scheme. assert_not_equals: got disallowed value "rgb(0, 0, 0)"
